+++ NEWS.md
@@ -1,8 +1,13 @@
 # healthyR.ai (development version)
 
+## Breaking Changes
+None
+
 ## New Features
 1. Fix #209 - Add function `hai_skewed_features()`
 2. Fix #210 - Export current internal functions.
+3. Fix #206 - Add functions `hai_scale_zscore_vec()` and `hai_scale_zscore_augment()` 
+and add `step_hai_scale_zscore()`
 
 ## Minor Fixes and Improvements
 1. Fix #208 - Enhance `hai_histogram_facet_plot()` It now returns a list output